+++ module/data/src/main/kotlin/io/github/siyual_park/data/aggregation/FetchContextProvider.kt
@@ -1,27 +1,17 @@
 package io.github.siyual_park.data.aggregation
 
-import com.google.common.cache.CacheBuilder
-import io.github.siyual_park.data.cache.InMemoryQueryStorage
-import io.github.siyual_park.data.cache.Pool
-import io.github.siyual_park.data.cache.PoolingNestedQueryStorage
-import io.github.siyual_park.data.cache.TransactionalQueryStorage
 import io.github.siyual_park.data.repository.QueryRepository
 import org.apache.commons.collections4.keyvalue.MultiKey
 import org.apache.commons.collections4.map.MultiKeyMap
 import kotlin.reflect.KClass
 
-class FetchContextProvider(
-    private val cacheBuilder: (() -> CacheBuilder<Any, Any>)
-) {
+class FetchContextProvider {
     private val contexts = MultiKeyMap<Any, FetchContext<*>>()
 
     fun <T : Any> get(repository: QueryRepository<T, *>, clazz: KClass<T>): FetchContext<T> {
         @Suppress("UNCHECKED_CAST")
         return contexts.getOrPut(MultiKey(repository, clazz)) {
-            val store = TransactionalQueryStorage(
-                PoolingNestedQueryStorage(Pool { InMemoryQueryStorage(clazz, cacheBuilder) })
-            )
-            FetchContext(store, repository, clazz)
+            FetchContext(repository, clazz)
         } as FetchContext<T>
     }
 }
