+++ module/data/src/main/kotlin/io/github/siyual_park/data/cache/CacheProvider.kt
@@ -1,22 +1,11 @@
 package io.github.siyual_park.data.cache
 
-import com.google.common.cache.Cache
 import com.google.common.cache.CacheBuilder
-import kotlinx.coroutines.sync.Mutex
-import kotlinx.coroutines.sync.withLock
-import java.util.Collections
-import java.util.WeakHashMap
 
 class CacheProvider<K : Any, T : Any>(
     cacheBuilder: CacheBuilder<Any, Any>,
 ) {
-    private val mutexes = Collections.synchronizedMap(WeakHashMap<K, Mutex>())
-    private val cache: Cache<K, T> = cacheBuilder
-        .removalListener<K, T> {
-            it.key?.let { id ->
-                mutexes.remove(id)
-            }
-        }.build()
+    private val cache = cacheBuilder.build<K, T>()
 
     private var updated = false
 
@@ -30,20 +19,12 @@ class CacheProvider<K : Any, T : Any>(
 
     suspend fun getIfPresent(key: K, value: suspend () -> T?): T? {
         return cache.getIfPresent(key) ?: run {
-            val mutex = mutexes.getOrPut(key) { Mutex() }
-            mutex.withLock {
-                val exists = cache.getIfPresent(key)
-                if (exists != null) {
-                    exists
-                } else {
-                    val newone = value()
-                    if (newone != null) {
-                        updated = true
-                        cache.put(key, newone)
-                    }
-                    newone
-                }
+            val newone = value()
+            if (newone != null) {
+                updated = true
+                cache.put(key, newone)
             }
+            newone
         }
     }
 
@@ -54,7 +35,6 @@ class CacheProvider<K : Any, T : Any>(
 
     suspend fun remove(key: K) {
         cache.invalidate(key)
-        mutexes.remove(key)
     }
 
     suspend fun entries(): Set<Pair<K, T>> {
@@ -65,7 +45,6 @@ class CacheProvider<K : Any, T : Any>(
         if (updated) {
             updated = false
             cache.invalidateAll()
-            mutexes.clear()
         }
     }
 }
