+++ module/data/src/main/kotlin/io/github/siyual_park/data/cache/PoolingNestedQueryStorage.kt
@@ -23,11 +23,11 @@ class PoolingNestedQueryStorage<T : Any>(
     }
 
     override suspend fun clear(entity: T) {
-        pool.used().forEach { it.clear(entity) }
+        pool.used().entries().forEach { it.clear(entity) }
     }
 
     override suspend fun clear() {
-        pool.used().forEach { it.clear() }
+        pool.used().entries().forEach { it.clear() }
         delegator.pop()?.let {
             pool.push(it)
         }
